2day/requirements.txt
uvicorn
pydantic
nameko
loguru
numpy
python-multipart
aiofiles
bson
fastapi
PyJWT==2.3.0
pymongo==4.1.1

